fixed AddIssueModal
