+++ src/components/organisms/AddIssueModal.js
@@ -4,33 +4,33 @@ import { PrimaryButton } from "../atoms/Button";
 import { useDispatch, useSelector } from "react-redux";
 import { closeModal, modalState } from "../../features/OpenModal";
 import { InputWindow, InputForm } from "../atoms/Text";
-import { addIssue, updateIssueInfo } from "../../features/IssueTableSlice";
+import { addIssue, Issue } from "../../features/IssueSlice";
 import { useRef } from "react";
 
-const AddIssueSection = styled.div`
+const MainSection = styled.div`
   max-width: 598px;
   margin: auto;
 `;
-const AddIssueTitle = styled.h2`
+const Header = styled.h2`
   padding: 0px 8px;
 `;
-const AddIssueBody = styled.div`
+const Body = styled.div`
   padding: 32px 0px 16px;
 `;
 
-const AddIssueInputSection = styled.div`
+const InputSection = styled.div`
   padding: 16px;
 `;
-const AddIssueFormHeaderText = styled.label`
+const FormText = styled.label`
   display: block;
   padding: 8px 0px;
 `;
 
-const AddIssueBlock = styled.div`
+const Description = styled.div`
   padding: 16px;
   min-height: 100px;
 `;
-const AddIssueModalFooter = styled.div`
+const Footer = styled.div`
   display: flex;
   -webkit-box-align: center;
   align-items: center;
@@ -38,20 +38,26 @@ const AddIssueModalFooter = styled.div`
   padding: 8px;
 `;
 
+function getDate() {
+  const today = new Date();
+  const month = String(today.getMonth() + 1).padStart(2, "0");
+  const day = String(today.getDate()).padStart(2, "0");
+  const year = String(today.getFullYear());
+  const date = month + "-" + day + "-" + year;
+  return date;
+}
+
 export default function AddIssueModal() {
   const inputTitleRef = useRef(null);
   const inputTextRef = useRef(null);
   const isOpen = useSelector(modalState);
   const dispatch = useDispatch();
-  const issues = useSelector(updateIssueInfo);
+  const issues = useSelector(Issue);
+
   function makeIssue() {
     const issueTitle = inputTitleRef.current.value;
     const issueText = inputTextRef.current.value;
-    const today = new Date();
-    const month = String(today.getMonth() + 1).padStart(2, "0");
-    const day = String(today.getDate()).padStart(2, "0");
-    const year = String(today.getFullYear());
-    const created = month + "-" + day + "-" + year;
+    const created = getDate();
     const num = issues.index + 1;
     const newIssue = {
       index: num,
@@ -71,30 +77,30 @@ export default function AddIssueModal() {
   return (
     <div>
       <ReactModal isOpen={isOpen}>
-        <AddIssueSection>
-          <AddIssueTitle>Issueを追加</AddIssueTitle>
-          <AddIssueBody>
-            <AddIssueInputSection>
-              <AddIssueFormHeaderText>タイトル</AddIssueFormHeaderText>
+        <MainSection>
+          <Header>Issueを追加</Header>
+          <Body>
+            <InputSection>
+              <FormText>タイトル</FormText>
               <InputWindow
                 type="input"
                 placeholder="タイトルを入力してください"
                 id="TitleForm"
                 ref={inputTitleRef}
               ></InputWindow>
-            </AddIssueInputSection>
-            <AddIssueInputSection>
-              <AddIssueFormHeaderText>説明</AddIssueFormHeaderText>
+            </InputSection>
+            <InputSection>
+              <FormText>説明</FormText>
               <InputForm
                 type="input"
                 placeholder="説明を入力してください"
                 id="TextForm"
                 ref={inputTextRef}
               ></InputForm>
-            </AddIssueInputSection>
-          </AddIssueBody>
-          <AddIssueBlock />
-          <AddIssueModalFooter>
+            </InputSection>
+          </Body>
+          <Description />
+          <Footer>
             <PrimaryButton isPrimary={true} onClick={() => makeIssue()}>
               作成
             </PrimaryButton>
@@ -104,8 +110,8 @@ export default function AddIssueModal() {
             >
               閉じる
             </PrimaryButton>
-          </AddIssueModalFooter>
-        </AddIssueSection>
+          </Footer>
+        </MainSection>
       </ReactModal>
     </div>
   );
