+++ src/components/pages/MainPage.js
@@ -1,5 +1,6 @@
 import React from "react";
-import { IssueHeaderSection, TabPage } from "./IssuePage";
+import IssueHeaderSection from "../organisms/IssueHeader";
+import { TabSection } from "../organisms/TabSection";
 import IssueTableSection from "../organisms/IssueTable";
 import { MainPage, MainTab, MainSection } from "../molecules/MainPagePart";
 import PullRequestPage from "./PullRequestPage";
@@ -19,12 +20,14 @@ export default function BodySection() {
     <>
       <MainPage>
         <MainTab>
-          <TabPage
+          <TabSection
+            //修正中 Childrenについて調べる
             children={"Issue"}
             isActive={isIssuePage}
             tabClick={issuePageClick}
           />
-          <TabPage
+          <TabSection
+            //修正中 Childrenについて調べる
             children={"Pull Request"}
             isActive={!isIssuePage}
             tabClick={pullRequestPageClick}
